First stab at a low contrast theme
I have issues when viewing brightly colored text on dark backgrounds. The text
appears to be popping off the screen and it leads to headaches.

This theme aims for much lower contrast, and will work best on a terminal with
a white background. It uses relatively fewer colors, and aims mostly for
shades of grey. Colors are used as accents to highlight some information, but
we should avoid pastels or any bright pallet colors.

I haven't tested out every possible segment, but for the ones I can easily
test this has so far been more suitable for me.