Create segments in parallel
Powerline being integrated directly in users shell, the perceived
performance and time to render and display the prompt is a critical
part of the program.

As discussed in #121, there are some integrations that requires
executing sub-programs and hence takes longer to render.

In order to improve the perceived speed, generate the segments data in
parallel, before sorting them to ensure the order the user required on
the command line.

After my testing, I am seeing that using modules
user,host,ssh,docker,kube,cwd,perms,git improves from 120 ms without the
update to 30 ms with the update.