venv option to only show indicator if name is long
Some virtualenvs have very long names, crowding the prompt. This change
adds an option to have the venv segment simply indicate that a virtualenv
is active if the name is longer than a given size. It defaults to 0,
which it considers unlimited, and therefore won't change behavior from
before if the user hasn't specified a limit.