add new flag -show-warnings
This flag could be used to toggle all warn-messages which are prompted
to stdout.

For example: if you use -shell-var flag and some variables aren't always
set (as expected) you will always see (for a very short time) the
prompt, that the variable isn't set. In combination with
-show-warnings=false your eyes are spared and you can be much more
focused on doing command line stuff ;-)

Signed-off-by: Mario Constanti <github@constanti.de>