feat(duration): add low precision duration setting
This brings in a new command line flag `-duration-low-precision` that
can be used to disable showing µs in the duration segment. This is
needed for shells like `fish` that only have millisecond precision or
for when people decide that they just don't want to have that extreme
level of precision.

Contributes to justjanne/powerline-go#307

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>